work with older versions of git
Before this commit we used `git worktree add --orphan` to create a new
`entomologist-data` branch, but older versions of git don't have that
option.  git 2.34 is known to break, 2.39 is known to work.

This commit switches to instead use a simpler `git worktree` invocation
to create a detached worktree, then `git switch --orphan` in the worktree
to create the orphan branch.  This is verified to work on git 2.30.